+++ javascript/index.js
@@ -1,9 +1,9 @@
-const { extractTbeHeader } = require('./src/functions/tbe-header-parser');
+const { stripHeader } = require('./src/functions/stripHeader');
 const errorMessages = require('./src/constants/error-messages');
 const warningMessages = require('./src/constants/warning-messages');
 
 module.exports = {
-  extractTbeHeader,
+  stripHeader,
   errorMessages,
   warningMessages
 };
